+++ paddlemix/__init__.py
@@ -17,5 +17,4 @@ from .datasets import *
 from .models import *
 from .optimization import *
 from .processors import *
-from .custom_ops import *
-
+from .triton_ops import *
